Extract Msf::DBManager::Import::Nexpose::Raw
MSP-11124
